// frontend/sdk/models/PlaylistItemKind.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sopranino API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const PlaylistItemKind = {
    Spotify: 'SPOTIFY',
    Imported: 'IMPORTED'
} as const;
export type PlaylistItemKind = typeof PlaylistItemKind[keyof typeof PlaylistItemKind];


export function instanceOfPlaylistItemKind(value: any): boolean {
    for (const key in PlaylistItemKind) {
        if (Object.prototype.hasOwnProperty.call(PlaylistItemKind, key)) {
            if (PlaylistItemKind[key as keyof typeof PlaylistItemKind] === value) {
                return true;
            }
        }
    }
    return false;
}

export function PlaylistItemKindFromJSON(json: any): PlaylistItemKind {
    return PlaylistItemKindFromJSONTyped(json, false);
}

export function PlaylistItemKindFromJSONTyped(json: any, ignoreDiscriminator: boolean): PlaylistItemKind {
    return json as PlaylistItemKind;
}

export function PlaylistItemKindToJSON(value?: PlaylistItemKind | null): any {
    return value as any;
}

export function PlaylistItemKindToJSONTyped(value: any, ignoreDiscriminator: boolean): PlaylistItemKind {
    return value as PlaylistItemKind;
}